References Oppevara/oppevara#110, deals with IE and Edge.
Changed the code to work with IE and Edge (IE needed more attention as
it does not support some of the newer APIs). Added a fix method that
would try to deal with attribute positioning issue. Does not try to fix
only vectors, but any <input> elements with more than one attribute.
This is just in case there are other cases like that. The code seems to
be working with other browsers without any issues. Safari is still
acting weird, but the cause is elsewhere.